To check weather a number is a palindrome or not
Palindrome is a word, phrase, or sequence that reads the same backwards as forwards. this program checks weather a number entered is a palindrome or not. it uses a simple while loop followed with a if else statement for printing weather a number is palindrome or not. it was fun creating this program