A  BASIC TIC TAK TOE GAME
Ive made a simple tick tak toe game with the pygame library. it uses the system graphics and sound to render the game. it uses an infinite while loop which is only satisfied ones one of the player wins. it also has some of the beautification done to it to make it look presentable. it was fun creating this project