A GAME OF HANGMAN
Hangman is a fun little word guessing game. The program uses lists which has different words stored into them. the random function is used to pick out a random word from the list. the player has to guess the correct word before the man is hanged. the program uses booleans, while and for loops, try and except , lists etc. it was a fun program to built.